requirements.txt
ccxt>=4.5.5
future>=1.0.0
pandas>=2.3.2
openpyxl>=3.1.5
certifi>=2025.1.31
aiohttp>=3.12.15
orjson>=3.11.3
python-telegram-bot>=22.5
python-dotenv>=1.0.0
pip >=25.3